+++ Firmware/SpoolJoin.cpp
@@ -1,4 +1,4 @@
-#include "AutoDeplete.h"
+#include "SpoolJoin.h"
 #include "eeprom.h"
 #include "Filament_sensor.h"
