+++ Firmware/Tcodes.cpp
@@ -1,5 +1,5 @@
 #include "Tcodes.h"
-#include "AutoDeplete.h"
+#include "SpoolJoin.h"
 #include "Marlin.h"
 #include "language.h"
 #include "messages.h"
